--- a/run_benchmarks.jl
+++ b/run_benchmarks.jl
@@ -12,9 +12,17 @@ for category in readdir()
     for test in readdir()
         endswith(test, ".jl") || continue
         @show test
-        result = open(deserialize, `$JULIAVER --project=. $test $RUNS SERIALIZE`)
-        #result=(value, times, bytes, gctime, gcstats)
-        println("median time: ", median(result.times))
+        result = (value=[], time=[], bytes=[], gctime=[], gcstats=[])
+        for _ in 1:RUNS
+            r = open(deserialize, `$JULIAVER --project=. $test SERIALIZE`)
+            push!(result.value,   r.value)
+            push!(result.time,    r.time)
+            push!(result.bytes,   r.bytes)
+            push!(result.gctime,  r.gctime)
+            push!(result.gcstats, r.gcstats)
+        end
+        #result=(value, time, bytes, gctime, gcstats)
+        println("median time: ", median(result.time))
         println("mean GC time: ", result.gctime)
     end
     cd("..")