--- a/run_one_benchmark.jl
+++ b/run_one_benchmark.jl
@@ -16,12 +16,19 @@ dir = joinpath(@__DIR__, path)
 cd(dir)
 @printf("Running Benchmark path = %s file = %s times = %d\n", path, file, RUNS)
 @show BENCH
-result = open(deserialize, `$JULIAVER --project=. $file $RUNS SERIALIZE`)
-(value,times,stats)= result
+value=[]
+times=[]
+stats=[]
+for _ in 1:RUNS
+   r = open(deserialize, `$JULIAVER --project=. $test SERIALIZE`)
+   push!(value, r.value)
+   push!(times, r.times)
+   push!(stats, r.stats)
+end
 @printf("run time: %0.0fms min, %0.0fms max %0.0fms median\n",
-   minimum(result.times) / 1_000_000,
-   maximum(result.times) / 1_000_000,
-   median(result.times)  / 1_000_000)
+   minimum(times) / 1_000_000,
+   maximum(times) / 1_000_000,
+   median(times)  / 1_000_000)
 
 time = map(gctime, stats)
 @printf("gc time: %0.0fms min, %0.0fms max, %0.0fms median\n",