now runs each benchmark in it's own process.