@chflood 2022-04-22
Before:
julia run_benchmarks.jl 3
category = "TimeZones"
test = "TimeZones.jl"
run time: 706ms min, 2335ms max 1100ms median
gc time: 0ms min, 1587ms max, 439ms median

After:
julia run_benchmarks.jl  3
category = "TimeZones"
test = "TimeZones.jl"
median time: 2.071309008
mean GC time: Any[1.384555556, 1.490022725, 1.481385896]

The before output is more informative.