@oscardssmith 2022-04-22
Oh, were you comparing to https://github.com/JuliaCI/GCBenchmarks/pull/8? If you merge that, I'll update this PR.